Change to float

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import asyncio
 from typing import AsyncGenerator
 
 
-async def async_generator() -> AsyncGenerator[int, None]:
+async def async_generator() -> AsyncGenerator[float, None]:
     ''' function that waits for 1 sec'''
     for i in range(10):
         await asyncio.sleep(1)
